break down handler class into different files
