+++ src/server/handlers/games/card.ts
@@ -1,8 +1,8 @@
 import { CardJitsuProgress } from "@server/game-logic/ninja-progress";
 import { Stamp } from "@server/game-logic/stamps";
 import { CardJitsu, NinjaPlayer, Sensei, WorldClient, WorldContext } from "@server/new-client";
-import { Handler, XtHandler } from "..";
 import { Handle } from "../handles";
+import { XtHandler } from "../xt";
 
 const handler = new XtHandler<WorldClient, WorldContext, ['world', 'penguin', 'card']>(['world', 'penguin', 'card']);
